+++ src/FSharpWrap.Tool/Filter.fs
@@ -2,51 +2,34 @@
 
 open System.Reflection
 
-[<RequireQualifiedAccess>]
-type AssemblyFiles =
-    | Exclude of Set<Path>
-    | Include of Set<Path>
+type FilterType<'Item when 'Item : comparison> =
+    | Exclude of Set<'Item>
+    | Include of Set<'Item>
     | All
 
-[<RequireQualifiedAccess>]
-type AssemblyNames =
-    | Exclude of Set<string>
-    | Include of Set<string>
-    | All
+    member this.Items =
+        match this with
+        | Exclude items
+        | Include items -> items
+        | All -> Set.empty
 
 type Filter =
-    { AssemblyFiles: AssemblyFiles
-      AssemblyNames: AssemblyNames
-      Namespaces: Set<Namespace> }
+    { Assemblies: FilterType<string>
+      Namespaces: FilterType<Namespace> }
 
     static member Empty =
-        { AssemblyFiles = AssemblyFiles.All
-          AssemblyNames = AssemblyNames.All
-          Namespaces = Set.empty }
+        { Assemblies = All
+          Namespaces = All }
 
 [<RequireQualifiedAccess>]
 module Filter =
-    let assemblyFiles { AssemblyFiles = files } =
-        match files with
-        | AssemblyFiles.Exclude files'
-        | AssemblyFiles.Include files' -> files'
-        | AssemblyFiles.All -> Set.empty
-
-    let assemblyNames { AssemblyNames = names } =
-        match names with
-        | AssemblyNames.Exclude names'
-        | AssemblyNames.Include names' -> names'
-        | AssemblyNames.All -> Set.empty
-
-    let private filterLocation files convert =
-        fun (assm: Assembly) ->
-            match Path.ofStr assm.Location with
-            | Some path -> Set.contains path files |> convert
-            | None -> true
+    let private all _ = true
 
     let assemblyIncluded filter =
-        match filter.AssemblyFiles with
-        | AssemblyFiles.Include inc -> filterLocation inc id
-        | AssemblyFiles.Exclude exc -> filterLocation exc not
-        | AssemblyFiles.All -> fun _ -> true
-        // TODO: Check assembly names
+        let found names (assm: Assembly) =
+            let name = assm.GetName().Name
+            Set.contains name names
+        match filter.Assemblies with
+        | Exclude exc -> found exc >> not
+        | Include inc -> found inc
+        | All -> all