+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Small performance improvements
 - Additional switches to specify which assemblies and namespaces to include in code generation
 ### Changed
-- Default assemblies that are included are now only NuGet packages or project references
+- All assemblies are included in code generation by default
 
 ## [0.3.0] - 2020-11-13
 ### Added